Split sheet.js into three modules (#96)
Separate the 911-line sheet.js into focused modules:
- sheet.js: orchestrator + shared render helpers (567 lines)
- sheet-screen.js: screen rendering (2-page DOM layout, 40 lines)
- sheet-print.js: print pagination (multi-column log, table
  splitting, anti-orphan logic, 295 lines)

Print functions receive the Sheet object as a parameter to avoid
circular imports. No functional changes — purely structural.

Also updates sw.js ASSETS and AGENTS.md.
